+++ OUDS/Core/Themes/Orange/Tests/Values/ComponentTokens/MockTheme/MockTheme+AllLinkComponentTokens.swift
@@ -27,20 +27,19 @@ final class MockThemeLinkComponentTokenProvider: OrangeThemeLinkComponentTokensP
     static let mockThemeLinkSize: DimensionRawToken = 118
     static let mockThemeLinkSpace: DimensionRawToken = 218
     static let mockThemeLinkColor = MultipleColorSemanticToken("#00FF00")
-    static let mockThemeExpandLinkIconStart = true // false in Orange Theme
-    static let mockThemeExpandLinkIconEnd = false // true in Orange Theme
 
     override init(sizes: AllSizeSemanticTokensProvider?,
                   colors: AllColorSemanticTokensProvider?,
-                  spaces: AllSpaceSemanticTokensProvider?)
+                  spaces: AllSpaceSemanticTokensProvider?,
+                  dimensions: AllDimensionSemanticTokensProvider?)
     {
-        super.init(sizes: sizes, colors: colors, spaces: spaces)
+        super.init(sizes: sizes, colors: colors, spaces: spaces, dimensions: dimensions)
     }
 
     // MARK: - Link component tokens
 
     override var sizeMinHeightDefault: SizeSemanticToken { Self.mockThemeLinkSize }
-    override var sizeMinWidthDefault: SizeSemanticToken { Self.mockThemeLinkSize }
+    override var sizeMinWidth: SizeSemanticToken { Self.mockThemeLinkSize }
     override var sizeMinHeightSmall: SizeSemanticToken { Self.mockThemeLinkSize }
     override var sizeMinWidthSmall: SizeSemanticToken { Self.mockThemeLinkSize }
     override var monoColorContentEnabled: MultipleColorSemanticToken { Self.mockThemeLinkColor }
@@ -49,7 +48,10 @@ final class MockThemeLinkComponentTokenProvider: OrangeThemeLinkComponentTokensP
     override var monoColorContentFocus: MultipleColorSemanticToken { Self.mockThemeLinkColor }
     override var monoColorContentDisabled: MultipleColorSemanticToken { Self.mockThemeLinkColor }
     override var spacePaddingInline: SpaceSemanticToken { Self.mockThemeLinkSpace }
-    override var spacePaddingBlock: SpaceSemanticToken { Self.mockThemeLinkSpace }
+    override var spacePaddingBlockDefault: SpaceSemanticToken { Self.mockThemeLinkSpace }
+    override var spacePaddingBlockSmall: SpaceSemanticToken { Self.mockThemeLinkSpace }
+    override var spacePaddingBlockCompactDensityDefault: SpaceSemanticToken { Self.mockThemeLinkSpace }
+    override var spacePaddingBlockCompactDensitySmall: SpaceSemanticToken { Self.mockThemeLinkSpace }
     override var spaceColumnGapIconDefault: SpaceSemanticToken { Self.mockThemeLinkSpace }
     override var spaceColumnGapIconSmall: SpaceSemanticToken { Self.mockThemeLinkSpace }
     override var spaceColumnGapChevronDefault: SpaceSemanticToken { Self.mockThemeLinkSpace }
@@ -64,11 +66,6 @@ final class MockThemeLinkComponentTokenProvider: OrangeThemeLinkComponentTokensP
     override var colorChevronHover: MultipleColorSemanticToken { Self.mockThemeLinkColor }
     override var colorChevronPressed: MultipleColorSemanticToken { Self.mockThemeLinkColor }
     override var colorChevronFocus: MultipleColorSemanticToken { Self.mockThemeLinkColor }
-
-    // MARK: - Expand link component tokens
-
-    override var iconStart: Bool { Self.mockThemeExpandLinkIconStart }
-    override var iconEnd: Bool { Self.mockThemeExpandLinkIconEnd }
 }
 
 // swiftlint:enable required_deinit