+++ OUDS/Core/ThemesContract/Sources/OUDSTheme/OUDSTheme.swift
@@ -97,7 +97,10 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
     // MARK: - Component tokens
     // Keep things alphabetically ordered
 
-    /// All components tokens related to alert component like `OUDSAlert`
+    /// All components tokens related to alert componens
+    public let accordion: AllAccordionComponentTokensProvider
+
+    /// All components tokens related to alert component like `OUDSAlertMessage` and `OUDSInlineAlert`
     public let alert: AllAlertComponentTokensProvider
 
     /// All components tokens related to bar components
@@ -121,19 +124,23 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
     /// All components tokens related to divider components like `OUDSHorizontalDivider` and `OUDSVerticalDivider`
     public let divider: AllDividerComponentTokensProvider
 
-    /// All component tokens related to control-item-layout-based components like `OUDSSwitch`, `OUDSRadioButtonItem` and `OUDSCheckboxItem`
-    public let controlItem: AllControlItemComponentTokensProvider
-
     /// All components tokens related to icons components
     public let icon: AllIconComponentTokensProvider
 
     /// All components tokens related to link components like `OUDSLink`
     public let link: AllLinkComponentTokensProvider
 
-    /// All components tokens related to pin code input components like `OUDSPinCodeInput`
+    /// All components tokens related to list item components like `OUDSListItem` and `OUDSSmallListItem`,
+    /// and also `OUDSSwitchItem`, `OUDSRadioButtonItem` and `OUDSCheckboxItem`
+    public let listItem: AllListItemComponentTokensProvider
+
+    /// All components tokens related to pin code input components
     public let pinCodeInput: AllPinCodeInputComponentTokensProvider
 
-    /// All components tokens related to quantity input components like `OUDSQuantityInput`
+    /// All components tokens related to progress indicator components
+    public let progressIndicator: AllProgressIndicatorComponentTokensProvider
+
+    /// All components tokens related to quantity input components
     public let quantityInput: AllQuantityInputComponentTokensProvider
 
     /// All components tokens related to checkboxes components like `OUDSRadioButton` and `OUDSRadioButtonItem`
@@ -142,7 +149,7 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
     /// All components tokens related to select input components like `OUDSSelectInput`
     public let selectInput: AllSelectInputComponentTokensProvider
 
-    /// All components tokens related to skeleto components like `OUDSSkeleton`
+    /// All components tokens related to skeleto components
     public let skeleton: AllSkeletonComponentTokensProvider
 
     /// All components tokens related to switch / toggle components like `OUDSSwitch`
@@ -159,6 +166,9 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
 
     /// All components tokens related to bullet list components like `OUDSTextInput`
     public let textInput: AllTextInputComponentTokensProvider
+
+    /// All components tokens related to typography components
+    public let typography: AllTypographyComponentTokensProvider
 
     // MARK: - Other elements
 
@@ -197,6 +207,7 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
     ///    - dimensions: All semantic tokens of dimensions
     ///    - sizes: All semantic tokens of sizes
     ///    - spaces: All semantic tokens of spaces
+    ///    - accordion: All component tokens for accordion
     ///    - alert: All component tokens for alert
     ///    - badge: All component tokens for badge
     ///    - bar: All component tokens for bar
@@ -204,11 +215,12 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
     ///    - button: All component tokens for button
     ///    - checkbox: All component tokens for checkbox
     ///    - chip: All component tokens for chip
-    ///    - controlItem: All component tokens for control item
     ///    - divider: All component tokens for divider
     ///    - icon: All component tokens for icon
     ///    - link: All component tokens for link
+    ///    - listItem: All component tokens for list item
     ///    - pinCodeInput: All component tokens for pin code input
+    ///    - progressIndicator: All component tokens for progress indicator
     ///    - quantityInput: All component tokens for quantity input
     ///    - radioButton: All component tokens for radio buttons
     ///    - selectInput: All component tokens for select input
@@ -218,6 +230,7 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
     ///    - inputTag: All component tokens for input tag
     ///    - textArea: All component tokens for text area
     ///    - textInput: All component tokens for text input
+    ///    - typography: All component tokens for typography
     ///    - name: The name of the theme, can be used for debugging for example
     ///    - resourcesBundle: The `Bundle` of the module containing the assets to load (e.g. icons of components, etc.)
     ///    - fontFamily: Set `nil` if system font to use, otherwise use the `FontFamilySemanticToken` you want to apply
@@ -235,6 +248,7 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
                 dimensions: AllDimensionSemanticTokensProvider,
                 sizes: AllSizeSemanticTokensProvider,
                 spaces: AllSpaceSemanticTokensProvider,
+                accordion: AllAccordionComponentTokensProvider,
                 alert: AllAlertComponentTokensProvider,
                 badge: AllBadgeComponentTokensProvider,
                 bar: AllBarComponentTokensProvider,
@@ -242,11 +256,12 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
                 button: AllButtonComponentTokensProvider,
                 checkbox: AllCheckboxComponentTokensProvider,
                 chip: AllChipComponentTokensProvider,
-                controlItem: AllControlItemComponentTokensProvider,
                 divider: AllDividerComponentTokensProvider,
                 icon: AllIconComponentTokensProvider,
                 link: AllLinkComponentTokensProvider,
+                listItem: AllListItemComponentTokensProvider,
                 pinCodeInput: AllPinCodeInputComponentTokensProvider,
+                progressIndicator: AllProgressIndicatorComponentTokensProvider,
                 quantityInput: AllQuantityInputComponentTokensProvider,
                 radioButton: AllRadioButtonComponentTokensProvider,
                 selectInput: AllSelectInputComponentTokensProvider,
@@ -256,6 +271,7 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
                 inputTag: AllInputTagComponentTokensProvider,
                 textArea: AllTextAreaComponentTokensProvider,
                 textInput: AllTextInputComponentTokensProvider,
+                typography: AllTypographyComponentTokensProvider,
                 resourcesBundle: Bundle,
                 name: String,
                 fontFamily: FontFamilySemanticToken? = nil,
@@ -278,6 +294,7 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
         self.spaces = spaces
 
         // Save component tokens providers
+        self.accordion = accordion
         self.alert = alert
         self.badge = badge
         self.bar = bar
@@ -285,11 +302,12 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
         self.bulletList = bulletList
         self.checkbox = checkbox
         self.chip = chip
-        self.controlItem = controlItem
         self.divider = divider
         self.icon = icon
         self.link = link
+        self.listItem = listItem
         self.pinCodeInput = pinCodeInput
+        self.progressIndicator = progressIndicator
         self.quantityInput = quantityInput
         self.radioButton = radioButton
         self.selectInput = selectInput
@@ -299,6 +317,7 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
         self.inputTag = inputTag
         self.textArea = textArea
         self.textInput = textInput
+        self.typography = typography
 
         // Load other configuration elements
         self.resourcesBundle = resourcesBundle